Lidt fejlfinding og oprydning

+++ assets/scripts/apps.js
@@ -1,24 +1,29 @@
 function deleteApp(button) {
-    let docId = button.getAttribute('data-docId')
+    if (confirm("Are you sure you want to delete this?")) {
+        let docId = button.getAttribute('data-docId')
 
-    let reqObj = {
-        DocId: docId
+        let reqObj = {
+            DocId: docId
+        }
+
+        let jsonString = JSON.stringify(reqObj)
+
+        fetch('/report/allApps/delete', {
+                method: 'delete',
+                headers: {
+                    'Accept': 'application/json, text/plain, */*',
+                    'Content-Type': 'application/json'
+                },
+                body: jsonString
+            })
+            .then(res => res.json())
+            .then(data => {
+                window.location.href = window.location.href;
+            });
+    } else {
+        return
     }
 
-    let jsonString = JSON.stringify(reqObj)
-
-    fetch('/report/allApps/delete', {
-            method: 'delete',
-            headers: {
-                'Accept': 'application/json, text/plain, */*',
-                'Content-Type': 'application/json'
-            },
-            body: jsonString
-        })
-        .then(res => res.json())
-        .then(data => {
-            window.location.href = window.location.href;
-        });
 
 }
